Refactor: Implementar arquitectura MVC (rutas, controladores, middlewares)
